2017.3.20
数据加滤波前备份

+++ USER/Others/Inc/Queue.h
@@ -4,17 +4,17 @@
 #include "Common.h"
 typedef struct Queue_{
   u8* BUF;      
-  u8  Size;    
-  u8  Head;           
-  u8  End;     
-	u8  Length;
+  u16  Size;    
+  u16  Head;           
+  u16  End;     
+	u16  Length;
   BOOL IsFull;     
   BOOL IsEmpty;  
 }Queue_;
 
 typedef struct Queue_OPS_
 {
-	Queue_ (*Init)(u8 Size,u8* P);
+	Queue_ (*Init)(u16 Size,u8* P);
 	BOOL (*Enqueue)(u8  byte,Queue_* Aim);
 	BOOL (*Dequeue)(u8* byte,Queue_* Aim);
